+++ libraries/base/Data/List/NonEmpty.hs
@@ -109,7 +109,9 @@ import           Data.Data           (Data)
 import           Data.Foldable       hiding (length, toList)
 import qualified Data.Foldable       as Foldable
 import           Data.Function       (on)
+import           Data.Functor.Classes (Eq1(..), Ord1(..), Read1(..), Show1(..))
 import qualified Data.List           as List
+import           Data.Monoid         ((<>))
 import           Data.Ord            (comparing)
 import qualified GHC.Exts            as Exts (IsList(..))
 import           GHC.Generics        (Generic, Generic1)
@@ -121,6 +123,27 @@ infixr 5 :|, <|
 -- @since 4.9.0.0
 data NonEmpty a = a :| [a]
   deriving ( Eq, Ord, Show, Read, Data, Generic, Generic1 )
+
+-- | @since 4.10.0.0
+instance Eq1 NonEmpty where
+  liftEq eq (a :| as) (b :| bs) = eq a b && liftEq eq as bs
+
+-- | @since 4.10.0.0
+instance Ord1 NonEmpty where
+  liftCompare cmp (a :| as) (b :| bs) = cmp a b <> liftCompare cmp as bs
+
+-- | @since 4.10.0.0
+instance Read1 NonEmpty where
+  liftReadsPrec rdP rdL p s = readParen (p > 5) (\s' -> do
+    (a, s'') <- rdP 6 s'
+    (":|", s''') <- lex s''
+    (as, s'''') <- rdL s'''
+    return (a :| as, s'''')) s
+
+-- | @since 4.10.0.0
+instance Show1 NonEmpty where
+  liftShowsPrec shwP shwL p (a :| as) = showParen (p > 5) $
+    shwP 6 a . showString " :| " . shwL as
 
 -- | @since 4.9.0.0
 instance Exts.IsList (NonEmpty a) where